server/requirements.txt
flask==2.0.0
Flask-JWT-Extended
python-dotenv
flask-mongoengine
flask-cors